# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Code Red Backend API
  version: 1.3.0
  description: >
    Production-style backend contract for the Code Red coaching institute platform.
    Canonical routes are versioned under `/api/v1`. This revision aligns the contract
    with the current frontend implementation, including authority-based navigation,
    inventory requests, payroll views, promotion flows, and faculty material publishing.
    Access control follows the frontend role families: student, faculty, parent,
    and administration (`admin`, `administration`, `director`, `superadmin`). Request
    payloads are validated with Pydantic, some background work is dispatched through
    Celery using Redis, and throttled endpoints may return HTTP 429.
servers:
  - url: http://localhost:3500
    description: Local Flask API server
  - url: http://localhost:3000
    description: Frontend dev server proxy, if configured
...
paths:
...
  /api/v1/attendance:
    get:
      tags: [Attendance]
      summary: List attendance records
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: studentId
          schema: { type: integer }
        - in: query
          name: subjectId
          schema: { type: integer }
        - in: query
          name: date
          schema: { type: string, format: date }
        - in: query
          name: class
          schema: { type: integer }
        - in: query
          name: section
          schema: { type: string }
        - in: query
          name: startDate
          schema: { type: string, format: date }
        - in: query
          name: endDate
          schema: { type: string, format: date }
      responses:
        '200':
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttendanceRecord'
    post:
      tags: [Attendance]
      summary: Create attendance sheet
      description: Accessible to faculty and administration.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceSubmission'
      responses:
        '201':
          description: Attendance created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountSuccessResponse'
    put:
      tags: [Attendance]
      summary: Update attendance sheet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceSubmission'
      responses:
        '200':
          description: Attendance updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountSuccessResponse'
  /api/v1/attendance/me/stats:
    get:
      tags: [Attendance]
      summary: Get current student attendance stats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Attendance statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceStats'
...
  /api/v1/parent/children/{childId}/attendance:
    get:
      tags: [Parent]
      summary: Get attendance summary for linked child
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChildIdPath'
      responses:
        '200':
          description: Attendance summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceStats'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  parameters:
...
    ChildIdPath:
      in: path
      name: childId
      required: true
      schema: { type: string }
...
  schemas:
...
    AttendanceRecord:
      type: object
      properties:
        id: { type: string }
        studentId: { type: string }
        subjectId: { type: string }
        date: { type: string, format: date }
        status: { type: string }
        markedBy: { type: string }
    AttendanceStats:
      type: object
      properties:
        total: { type: integer }
        present: { type: integer }
        absent: { type: integer }
        percentage: { type: number }
...
    CountSuccessResponse:
      type: object
      properties:
        success: { type: boolean }
        count: { type: integer }
...
    AttendanceSubmission:
      type: object
      required: [date, class, records]
      properties:
        date: { type: string, format: date }
        class: { type: string }
        section: { type: string, nullable: true }
        records:
          type: array
          items:
            type: object
            required: [studentId, status]
            properties:
              studentId: { type: string }
              subjectId: { type: string, nullable: true }
              status: { type: string }
...